pyproject.toml
[project]
name = "toolkitx"
version = "0.0.1"
description = "A personal Python toolkit for common tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diskcache>=5.6.3",
    "py-transgpt>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-dotenv",
]

[project.scripts]
hello = "toolkitx.hello:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["toolkitx"]

[project.urls]
"Homepage" = "https://github.com/ider-zh/toolkitx"
"Bug Tracker" = "https://github.com/ider-zh/toolkitx/issues"
